Move logs utils to their separate folder under Logs
