+++ public/app/features/explore/Logs/Logs.tsx
@@ -45,11 +45,11 @@ import { ExploreId } from 'app/types/explore';
 
 import { LogRows } from '../../logs/components/LogRows';
 import { LogRowContextModal } from '../../logs/components/log-context/LogRowContextModal';
-import { SETTINGS_KEYS } from '../utils/logs';
 
 import { LogsMetaRow } from './LogsMetaRow';
 import LogsNavigation from './LogsNavigation';
 import { LogsVolumePanelList } from './LogsVolumePanelList';
+import { SETTINGS_KEYS } from './utils/logs';
 
 interface Props extends Themeable2 {
   width: number;
