+++ public/app/features/explore/Logs/LogsSamplePanel.tsx
@@ -19,7 +19,8 @@ import store from 'app/core/store';
 
 import { LogRows } from '../../logs/components/LogRows';
 import { SupplementaryResultError } from '../SupplementaryResultError';
-import { SETTINGS_KEYS } from '../utils/logs';
+
+import { SETTINGS_KEYS } from './utils/logs';
 
 type Props = {
   queryResponse: DataQueryResponse | undefined;
