+++ public/app/features/explore/Logs/utils/LogsCrossFadeTransition.tsx
